Changed sout to log.info

+++ src/main/java/com/andersen/internetShop/ProductRepository.java
@@ -24,7 +24,7 @@ public class ProductRepository {
         return products;
     }
 
-    public Product getById(Long id) {
+    public Product getById(Integer id) {
         return products.stream()
                 .filter(product -> product.getId().equals(id))
                 .findFirst()
